Топ, профиль и прочее

+++ src/main/java/ruslan2570/bobrcoin/entity/BobrEntity.java
@@ -27,7 +27,7 @@ public class BobrEntity {
     @JoinColumn(name = "type_id", nullable = false)
     private BobrTypeEntity bobrType;
 
-    private int lifetime;
+    private Long lifetime;
 
     private BigDecimal earnedForLife;
